% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aa_count.R
\name{aa_count}
\alias{aa_count}
\title{Plotting amino-acid count}
\usage{
aa_count(seq)
}
\arguments{
\item{seq}{an amino acid sequence in string format}
}
\value{
a bar plot with the amino acid count
}
\description{
Plotting amino-acid count
}
\examples{

seq1 <- "AKLPYNILSDFHVNSTRLGE"
print(aa_count(seq1))


}
