% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{DNA_codons}
\alias{DNA_codons}
\title{DNA Codons}
\usage{
DNA_codons(DNA_sequence, start = 1)
}
\arguments{
\item{DNA_sequence}{String of bases A, T, G, C or U.}

\item{start}{Position to start splitting string into codons.}
}
\value{
Strings of codons
}
\description{
DNA Codons
}
\examples{
DNA1 = 'ATGGCCAATGGGCCATGCT'
DNA_codons(DNA1)
}
