% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_to_rna.R
\name{dna_to_rna}
\alias{dna_to_rna}
\title{Transcribe DNA to RNA}
\usage{
dna_to_rna(dna_sequence)
}
\arguments{
\item{dna_sequence}{A character string representing a DNA sequence (e.g., "ATGCTTAC").}
}
\value{
A character string representing the transcribed RNA sequence, with all "T" characters replaced by "U".
}
\description{
This function takes a DNA sequence and transcribes it to an RNA sequence by replacing each occurrence of thymine ("T") with uracil ("U").
}
\examples{
# Example DNA sequence transcription
dna_sequence <- 'ATGCTTAC'
rna_sequence <- dna_to_rna(dna_sequence)
print(rna_sequence) # Expected output: 'AUGCUUAC'
}
