% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_aa_seq.R
\name{gen_aa_seq}
\alias{gen_aa_seq}
\title{Generate aminoacid sequence}
\usage{
gen_aa_seq(codons)
}
\arguments{
\item{codons}{this is just the triplet from the mRNA sequence}
}
\value{
paste_aa This function just joins amino acids without space in between
}
\description{
Generate aminoacid sequence
}
\examples{
codons <-  c("AAT", "TGC", "GGC")
aa_seq <- gen_aa_seq(codons)
print(aa_seq)
}
