% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_dna_seq.R
\name{gen_dna_seq}
\alias{gen_dna_seq}
\title{Generate a Random DNA Sequence}
\usage{
gen_dna_seq(length)
}
\arguments{
\item{length}{An integer specifying the length of the DNA sequence.}
}
\value{
A character string representing a DNA sequence of the specified length.
}
\description{
This function generates a random DNA sequence of a specific length.
}
\examples{
# Generate a random DNA sequence of length 10
length = 10
dna_seq <- gen_dna_seq(length)
print(dna_seq)
}
