% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_seq.R
\name{gen_seq}
\alias{gen_seq}
\title{Generate DNA sequence}
\usage{
gen_seq(seq_length)
}
\arguments{
\item{seq_length}{Length of sequence}
}
\value{
start start index
}
\description{
Generate DNA sequence
}
\examples{
gen_seq(7)
}
