% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dna.R
\name{generate_dna}
\alias{generate_dna}
\title{Generation of randomly sampled DNA sequence of a certain length}
\usage{
generate_dna(dna_length)
}
\arguments{
\item{dna_length}{The length of the DNA sequence you want to generate.
Enter as double or character.}
}
\value{
Randomly generated DNA sequence of length ´dna_length´ as character
}
\description{
Generation of randomly sampled DNA sequence of a certain length
}
\examples{

generate_dna(12)

generate_dna("21")
}
