% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_random_dna_sequence.R
\name{generate_random_dna_sequence}
\alias{generate_random_dna_sequence}
\title{Title
Generate random DNA sequence}
\usage{
generate_random_dna_sequence(length)
}
\arguments{
\item{length}{the length that the generated DNA seq should be}
}
\value{
returns a random generated DNA sequence
}
\description{
Title
Generate random DNA sequence
}
\examples{
length = 1000
generate_random_dna_sequence(length)

}
