% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genetic_code}
\alias{genetic_code}
\title{genetic_code}
\format{
An object of class \code{character} of length 64.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}
}
\usage{
genetic_code
}
\description{
This object contains the genetic code, that is, the equivalence between a codon and the resulting amino acid after biological translation.
}
\keyword{datasets}
