% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F3_make_codons.R
\name{make_codons}
\alias{make_codons}
\title{Dividing an mRNA string into 3mers (codons)}
\usage{
make_codons(mRNA_seq, start = 1)
}
\arguments{
\item{start}{the translational starting point. This will be the first letter of the first codon returned. Default is 1}

\item{mRNA}{a string of the letters A, U, C and G}
}
\value{
a vector of strings, each string having length 3.
}
\description{
Dividing an mRNA string into 3mers (codons)
}
