% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F2_transcribe_to_mRNA.R
\name{mRNA}
\alias{mRNA}
\title{Make mRNA sequence from DNA sequence}
\usage{
mRNA(dna_seq)
}
\arguments{
\item{dna_seq}{a string of the letters A, T, C and G}
}
\value{
the same string of letters as input, but with U substituting T
}
\description{
Make mRNA sequence from DNA sequence
}
