% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_amino_acid_composition.R
\name{plot_amino_acid_composition}
\alias{plot_amino_acid_composition}
\title{Plot amino acid composition}
\usage{
plot_amino_acid_composition(protein_sequence)
}
\arguments{
\item{protein_sequence}{Sequence of amino acids}
}
\value{
A bar plot
}
\description{
Creates a bar plot showing the counts of unique amino acids
from an one-letter protein sequence. This function uses stringr
for character processing and ggplot2 for visualization.
}
