% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_amino_acid_counts.R
\name{plot_amino_acid_counts}
\alias{plot_amino_acid_counts}
\title{Title
Plot Amino Acid Counts}
\usage{
plot_amino_acid_counts(amino_acid_sequence)
}
\arguments{
\item{amino_acid_sequence}{input is a amino acid seq}
}
\value{
plot of amino acid counts
}
\description{
Title
Plot Amino Acid Counts
}
\examples{
amino_seq = 'AAACGAGGCUAA'
plot_amino_acid_counts(amino_seq)

}
