% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prot_seq.R
\name{prot_seq}
\alias{prot_seq}
\title{Amino acid from RNA}
\usage{
prot_seq(codons)
}
\arguments{
\item{codons}{Takes a set of codons}
}
\value{
and merge them into one amino acid.
}
\description{
Amino acid from RNA
}
\examples{
prot_seq("AUA")
}
