% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_codon.R
\name{split_codon}
\alias{split_codon}
\title{Split DNA sequence into codons}
\usage{
split_codon(dna_seq, start = 1)
}
\arguments{
\item{dna_seq}{String of the input DNA sequence}

\item{start}{Starting codon position}
}
\value{
Vector of codons extracted from input string
}
\description{
Split DNA sequence into codons
}
\examples{
split_codon("ATGCGTGAC")

dna <-("TATGCGCCAAA")
pos=2
split_codon(dna, pos)
}
