% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f3.R
\name{split_to_codons}
\alias{split_to_codons}
\title{Split a Sequence into 3-Character Codons}
\usage{
split_to_codons(seq, start = 1)
}
\arguments{
\item{seq}{sequence to split}

\item{start}{start index}
}
\value{
a list of codons
}
\description{
Split a Sequence into 3-Character Codons
}
\examples{
split_to_codons("AGUGCAG")
}
