% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substi_T_U.R
\name{substi_T_U}
\alias{substi_T_U}
\title{Substitute Thymine with Uracile}
\usage{
substi_T_U(substi_T_U)
}
\arguments{
\item{substi_T_U}{a DNA sequence}
}
\value{
a RNA sequence
}
\description{
Substitute Thymine with Uracile
}
\examples{
substi_T_U("ATTCGGGTT")
}
