% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription.R
\name{transcribe}
\alias{transcribe}
\title{Transcription}
\usage{
transcribe(dna_seq)
}
\arguments{
\item{dna_seq}{The input is the DNA sequence}
}
\value{
The output should be the RNA sequence
}
\description{
Transcription
}
\examples{
transcript('AGCTAACTG')
}
