% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{translation}
\alias{translation}
\title{translation}
\usage{
translation(codons)
}
\arguments{
\item{codons}{a list of RNA codons}
}
\value{
a string of Amino Acids
}
\description{
translation
}
\examples{
translation(c("UAU","UUU","GAG"))
}
