% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F5_visualize_aa_counts.R
\name{visualize_aa_counts}
\alias{visualize_aa_counts}
\title{Visualize the composition of amino acids in a given amino acid sequence}
\usage{
visualize_aa_counts(aa_code)
}
\arguments{
\item{aa_code}{a string where each letter denotes an amino acid}
}
\value{
a ggplot2 bar plot visualization of the counts of occurences of each amino acid in the input sequence
}
\description{
Visualize the composition of amino acids in a given amino acid sequence
}
