% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F1_generate_random_DNA_seq.R
\name{generate_dna}
\alias{generate_dna}
\title{Generate random DNA sequence}
\usage{
generate_dna(dna_length)
}
\arguments{
\item{dna_length}{user specified input of DNA sequence, integer}
}
\value{
a randomly generated DNA sequence consisting of the bases A, T, C and G. Returned as a string
}
\description{
Generate random DNA sequence
}
