% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRNA.R
\name{mRNA}
\alias{mRNA}
\title{Convert DNA sequence into mRNA sequence}
\usage{
mRNA(dna_seq)
}
\arguments{
\item{dna_seq}{DNA sequence as a string}
}
\value{
mRNA sequence as a string
}
\description{
Convert DNA sequence into mRNA sequence
}
\examples{

mRNA("ATGCTTTAAGTA")
}
