% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_codon.R
\name{split_codon}
\alias{split_codon}
\title{Split DNA sequence into codons}
\usage{
split_codon(rna_seq, start = 1)
}
\arguments{
\item{start}{Starting codon position}

\item{dna_seq}{String of the input RNA sequence}
}
\value{
Vector of codons extracted from input string
}
\description{
Split DNA sequence into codons
}
\examples{
split_codon("AUGCGUGAC")

rna <-("UAUGCGCCAAA")
pos=2
split_codon(rna, pos)
}
