% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F4_translate_to_protein.R
\name{translation}
\alias{translation}
\title{Translation of a vector of codons into their corresponding amino acids based on the genetic code}
\usage{
translation(codons)
}
\arguments{
\item{codons}{a vector of strings, each of which are 3mers of the letters A, U, C and G}
}
\value{
a single string, with each letter symbolizing an amino acid
}
\description{
Translation of a vector of codons into their corresponding amino acids based on the genetic code
}
